Re-arranged some things, and added capability to cache task specific FT flat datasets.
